v.2.1.0.79: OwnerDetails - modified v_OwnerDetails to include inactive Owners so the list can be filtered by Active, Inactive or All. The default filter is Active Owners.

+++ DXApplication2/ViewModels/InvoiceViewModel.cs
@@ -220,7 +220,7 @@
                     Invoice invoice = new Invoice();
                     invoice.OwnerID = o.OwnerID;
                     invoice.MailTo = o.MailTo;
-                    invoice.Balance = (from y in dc.v_OwnerDetails
+                    invoice.Balance = (decimal)(from y in dc.v_OwnerDetails
                                        where y.OwnerID == o.OwnerID
                                        select y.Balance).FirstOrDefault();
